rpg-game/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.6.4
